Added java doc in BadgeActionMenuItemView Implemented setBadgePadding and getBadgePadding methods in BadgeActionMenuItemView

+++ library/src/main/java/com/badgeview/BadgeActionView.java
@@ -153,7 +153,7 @@ public class BadgeActionView extends BadgeActionMenuItemView {
      * @deprecated Use {@link BadgeActionMenuItemView#getBadgeDuration()} instead.
      */
     @Deprecated
-    public int getDuration() {
+    public long getDuration() {
         return super.getBadgeDuration();
     }
